refactor: move StateMachine etc to sdk
