+++ packages/sdk/src/utils/index.ts
@@ -8,3 +8,11 @@ export {
 export { Deferred } from "./Deferred";
 export { sql } from "./sqlTemplate";
 export { withTimeoutWarning, withTimeoutCallback } from "./timeout";
+export {
+  type TrackableState,
+  trackableState,
+  type StateMachine,
+  type StatusMachine,
+  stateMachine,
+  statusMachine,
+} from "./TrackableState";
